Move `SolarCalculator` into Libraries
