+++ Sming/Libraries/SolarCalculator/README.rst
@@ -1,0 +1,10 @@
+Solar Calculator
+================
+
+Arduino library to support calculation of sunrise / sunset times
+
+See :sample:`SystemClock_NTP` for example usage.
+
+This is straight port of the code used by https://www.esrl.noaa.gov/gmd/grad/solcalc/
+
+Javascript reference: https://www.esrl.noaa.gov/gmd/grad/solcalc/main.js
