+++ Sming/Libraries/SolarCalculator/include/SolarCalculator.h
