+++ samples/SystemClock_NTP/component.mk
@@ -1,1 +1,2 @@
 DISABLE_SPIFFS = 1
+ARDUINO_LIBRARIES = SolarCalculator
